requirements.txt
# requirements.txt
pycryptodome>=3.20.0
colorama>=0.4.6
tqdm>=4.66.0
pydantic>=2.0.0